MultiplesOfPi.jl

MultiplesOfPi.Pi — Constant

Pi

The number PiExpTimes(1, 1), numerically equivalent to pi. Using Pi instead of pi often produces results that avoid floating-point inaccuracies.

Examples

julia> sin(Pi)
0.0

julia> sin(π)
1.2246467991473532e-16

julia> exp(im*Pi) + 1 == 0
true

julia> exp(im*π) + 1 == 0
false

julia> (2//3)Pi + Pi == (5//3)Pi
true

julia> (2//3)π + π == (5//3)π
false

MultiplesOfPi.PiExpTimes — Type

PiExpTimes(x, n)

Construct a number that behaves as x*π^n for a real x and an integer n

Examples

julia> PiExpTimes(2, 3)
2Pi^3

See also: Pi